requirements.txt
autopep8==1.5.3
certifi==2020.4.5.1
chardet==3.0.4
cx-Oracle==7.3.0
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.19.1
openpyxl==3.0.3
pandas==1.1.1
protobuf==3.12.2
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
xlrd==1.2.0
XlsxWriter==1.3.5
